+++ samples/implementing-yourown-provider/InMemoryCache.Sample.ps1
@@ -28,7 +28,7 @@ function Invoke-Main {
         Initialize-ExpressionCache -Providers $inMemSpec -AppName "InMemoryCache.Sample" -ReplaceProviders:$SoleProvider
 
         # 3) Sanity check: ensure our provider is present
-        $registered = Get-ExpressionCacheProvider -ProviderName $ProviderName -NoFallback
+        $registered = Get-ExpressionCacheProvider -ProviderName $ProviderName -ErrorAction SilentlyContinue
         if (-not $registered) {
             throw "Provider '$ProviderName' was not registered. Initialization failed."
         }