+++ src/Utilities/Initialize-ProviderState.ps1
@@ -1,7 +1,7 @@
 function Initialize-ProviderState {
     param($Provider)
     if ($null -eq $Provider.State) {
-        With-ProviderLock $Provider {
+        Invoke-ProviderLockedOperation -Provider $Provider {
             if ($null -eq $Provider.State) {
                 $Provider.State = New-Object 'System.Collections.Concurrent.ConcurrentDictionary[string, object]'
             }